Extract `ActionOptionItem`
